Move list to separate file for testing
